requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
click==8.0.1
Django==3.2.5
libpcap==1.10.0b15
mypy-extensions==0.4.3
pathspec==0.9.0
pytz==2021.1
regex==2021.7.6
sqlparse==0.4.1
tomli==1.2.0
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
click==8.0.1
Django==3.2.5
libpcap==1.10.0b15
mypy-extensions==0.4.3
paho-mqtt==1.5.1
pathspec==0.9.0
pytz==2021.1
regex==2021.7.6
sqlparse==0.4.1
tomli==1.2.0
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
click==8.0.1
Django==3.2.5
libpcap==1.10.0b15
mypy-extensions==0.4.3
paho-mqtt==1.5.1
pathspec==0.9.0
psycopg2-binary==2.9.1
pytz==2021.1
regex==2021.7.6
sqlparse==0.4.1
tomli==1.2.0
